+++ Cartify.Domain/Entities/Category.cs
@@ -2,7 +2,7 @@ using Cartify.Domain.Entities.JoinEntities;
 
 namespace Cartify.Domain.Entities;
 
-public class Category : BaseEntity<long>
+public class Category : BaseEntity<int>
 {
     public required string Name { get; set; }
     public required string Slug { get; set; }
